# Swagger.yaml
openapi: 3.1.0
info:
  title: Auto Escola N-116
  description: API de Back-End da Auto Escola N-116.
  contact:
    name: SENAI
    email: carlos.moliveira@sp.senai.br
  license:
    name: Apache 2.0
    url: 'http://autoescolan116.com.br/api/license'
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
paths:
  /usuarios:
    get:
      tags:
        - usuario-controller
      operationId: listar
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageDadosListagemUsuario'
    put:
      tags:
        - usuario-controller
      operationId: atualizar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosAtualizacaoUsuario'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoUsuario'
    post:
      tags:
        - usuario-controller
      operationId: cadastrar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosCadastroUsuario'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoUsuario'
    patch:
      tags:
        - usuario-controller
      operationId: atualizarSenha
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosAtualizacaoSenha'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosSucesso'
  /instrutores:
    get:
      tags:
        - instrutor-controller
      operationId: listar_1
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageDadosListagemInstrutor'
    put:
      tags:
        - instrutor-controller
      operationId: atualizar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosAtualizacaoInstrutor'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoInstrutor'
    post:
      tags:
        - instrutor-controller
      operationId: cadastrar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosCadastroInstrutor'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoInstrutor'
  /alunos:
    get:
      tags:
        - aluno-controller
      operationId: listar_2
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageDadosListagemAluno'
    put:
      tags:
        - aluno-controller
      operationId: atualizar_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosAtualizacaoAluno'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoAluno'
    post:
      tags:
        - aluno-controller
      operationId: cadastrar_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosCadastroAluno'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoAluno'
  /login:
    post:
      tags:
        - login-controller
      operationId: logar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosLogin'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosTokenJWT'
  /instrucoes:
    post:
      tags:
        - instrucao-controller
      operationId: agendar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DadosAgendamento'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoAgendamento'
  '/usuarios/{id}':
    get:
      tags:
        - usuario-controller
      operationId: detalhar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoUsuario'
    delete:
      tags:
        - usuario-controller
      operationId: excluir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  '/instrutores/{id}':
    get:
      tags:
        - instrutor-controller
      operationId: detalhar_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoInstrutor'
    delete:
      tags:
        - instrutor-controller
      operationId: excluir_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /health_check:
    get:
      tags:
        - health-check-controller
      operationId: healthCheck
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  '/alunos/{id}':
    get:
      tags:
        - aluno-controller
      operationId: detalhar_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DadosDetalhamentoAluno'
    delete:
      tags:
        - aluno-controller
      operationId: excluir_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    DadosAtualizacaoUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        perfil:
          type: string
          enum:
            - ADMIN
            - USER
      required:
        - id
    DadosDetalhamentoUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        ativo:
          type: boolean
        perfil:
          type: string
          enum:
            - ADMIN
            - USER
    DadosAtualizacaoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
          format: email
        telefone:
          type: string
          pattern: '\(\d{2}\) \d{5}-\d{4}'
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: '#/components/schemas/DadosEndereco'
      required:
        - id
    DadosEndereco:
      type: object
      properties:
        logradouro:
          type: string
          minLength: 1
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
          minLength: 1
        cidade:
          type: string
          minLength: 1
        uf:
          type: string
          minLength: 1
          pattern: '[A-Z]{2}'
        cep:
          type: string
          minLength: 1
          pattern: '[0-9]{8}'
      required:
        - bairro
        - cep
        - cidade
        - logradouro
        - uf
    DadosDetalhamentoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cnh:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: '#/components/schemas/DadosEndereco'
        ativo:
          type: boolean
    DadosAtualizacaoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
          format: email
        telefone:
          type: string
          pattern: '\(\d{2}\) \d{5}-\d{4}'
        endereco:
          $ref: '#/components/schemas/DadosEndereco'
      required:
        - id
    DadosDetalhamentoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cpf:
          type: string
        endereco:
          $ref: '#/components/schemas/DadosEndereco'
        ativo:
          type: boolean
    DadosCadastroUsuario:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
        perfil:
          type: string
          enum:
            - ADMIN
            - USER
      required:
        - login
        - perfil
        - senha
    DadosLogin:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
      required:
        - login
        - senha
    DadosTokenJWT:
      type: object
      properties:
        tokenJWT:
          type: string
    DadosCadastroInstrutor:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
          pattern: '\(\d{2}\) \d{5}-\d{4}'
        cnh:
          type: string
          minLength: 1
          pattern: '\d{9,11}'
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: '#/components/schemas/DadosEndereco'
      required:
        - cnh
        - email
        - especialidade
        - nome
        - telefone
    DadosAgendamento:
      type: object
      properties:
        idAluno:
          type: integer
          format: int64
        idInstrutor:
          type: integer
          format: int64
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        data:
          type: string
          format: date-time
      required:
        - data
        - idAluno
    DadosDetalhamentoAgendamento:
      type: object
      properties:
        id:
          type: integer
          format: int64
        aluno:
          type: string
        instrutor:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        data:
          type: string
          format: date-time
    DadosCadastroAluno:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
          pattern: '\(\d{2}\) \d{5}-\d{4}'
        cpf:
          type: string
          minLength: 1
          pattern: '\d{9,11}'
        endereco:
          $ref: '#/components/schemas/DadosEndereco'
      required:
        - cpf
        - email
        - nome
        - telefone
    DadosAtualizacaoSenha:
      type: object
      properties:
        id:
          type: integer
          format: int64
        senhaAntiga:
          type: string
          minLength: 1
        senhaNova:
          type: string
          minLength: 1
          pattern: '\d{8}'
      required:
        - id
        - senhaAntiga
        - senhaNova
    DadosSucesso:
      type: object
      properties:
        message:
          type: string
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    DadosListagemUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        perfil:
          type: string
          enum:
            - ADMIN
            - USER
    PageDadosListagemUsuario:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/DadosListagemUsuario'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        unpaged:
          type: boolean
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    DadosListagemInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
    PageDadosListagemInstrutor:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/DadosListagemInstrutor'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    DadosListagemAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
    PageDadosListagemAluno:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/DadosListagemAluno'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT