# src/main/java/br/com/senain116/autoescolan116/adapter/in/controller/request/endereco/DadosEndereco.java
package br.com.senain116.autoescolan116.adapter.in.controller.request.endereco;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;

public record DadosEndereco(
        @NotBlank
        String logradouro,
        String numero,
        String complemento,

        @NotBlank
        String bairro,

        @NotBlank
        String cidade,

        @NotBlank
        @Pattern(regexp = "[A-Z]{2}")
        String uf,

        @NotBlank
        @Pattern(regexp = "[0-9]{8}")//Segunda maneira de padrão numérico
        String cep) {
}